function that reverses the content of an array of integers.

+++ pointers_arrays_strings/4-rev_array.c
@@ -10,7 +10,7 @@ void reverse_array(int *a, int n)
 {
 	int *l_ptr;
 	int *r_ptr;
-	unsigned int temp;
+	long int temp;
 
 	l_ptr = a;
 	r_ptr = &a[n - 1];
